Fix dependency issues.
 * Move client-related stuff into helios-client, e.g. AuthProvider
 * Move client auth plugin related code into helios-tools
 * Remove helios-transport module, it's no longer needed (no more
   circular dependency issues)
