+++ helios-authentication/src/main/java/com/spotify/helios/auth/AuthenticationPluginLoader.java
@@ -18,15 +18,11 @@
 package com.spotify.helios.auth;
 
 import com.google.common.base.Preconditions;
-import com.google.common.collect.ImmutableList;
-import com.google.common.collect.Iterators;
-import com.google.common.collect.Lists;
 
 import java.net.MalformedURLException;
 import java.net.URL;
 import java.net.URLClassLoader;
 import java.nio.file.Path;
-import java.util.List;
 import java.util.ServiceLoader;
 
 public class AuthenticationPluginLoader {
@@ -57,18 +53,6 @@ public class AuthenticationPluginLoader {
     } catch (MalformedURLException e) {
       throw new RuntimeException("Failed to load plugin jar " + plugin, e);
     }
-  }
-
-  public static List<AuthenticationPlugin> loadAll() {
-    return loadAll(null);
-  }
-
-  public static List<AuthenticationPlugin> loadAll(final Path pluginPath) {
-    final ServiceLoader<AuthenticationPlugin> loader =
-        serviceLoaderForPath(pluginPath);
-    final List<AuthenticationPlugin> plugins = Lists.newArrayList();
-    Iterators.addAll(plugins, loader.iterator());
-    return ImmutableList.copyOf(plugins);
   }
 
   private static ServiceLoader<AuthenticationPlugin> serviceLoaderForPath(final Path pluginPath) {
