+++ helios-authentication/src/main/java/com/spotify/helios/auth/basic/BasicAuthenticationPlugin.java
@@ -19,9 +19,7 @@ package com.spotify.helios.auth.basic;
 
 import com.google.auto.service.AutoService;
 
-import com.spotify.helios.auth.AuthProvider;
 import com.spotify.helios.auth.AuthenticationPlugin;
-import com.spotify.helios.transport.RequestDispatcher;
 
 import io.dropwizard.auth.basic.BasicCredentials;
 
@@ -37,27 +35,5 @@ public class BasicAuthenticationPlugin implements AuthenticationPlugin<BasicCred
   @Override
   public ServerAuthentication<BasicCredentials> serverAuthentication() {
     return new BasicServerAuthentication();
-  }
-
-  @Override
-  public ClientAuthentication clientAuthentication() {
-    final String username = System.getenv("AUTH_BASIC_USERNAME");
-    final String password = System.getenv("AUTH_BASIC_PASSWORD");
-
-    if (username == null || password == null) {
-      return null;
-    } else {
-      return new ClientAuthentication() {
-        @Override
-        public AuthProvider.Factory authProviderFactory() {
-          return new AuthProvider.Factory() {
-            @Override
-            public AuthProvider create(final RequestDispatcher requestDispatcher) {
-              return new BasicAuthProvider(username, password);
-            }
-          };
-        }
-      };
-    }
   }
 }
