+++ helios-client/src/main/java/com/spotify/helios/client/DefaultRequestDispatcher.java
@@ -29,9 +29,6 @@ import com.google.common.util.concurrent.ListeningExecutorService;
 import com.fasterxml.jackson.core.type.TypeReference;
 import com.spotify.helios.common.HeliosException;
 import com.spotify.helios.common.Json;
-import com.spotify.helios.transport.HeliosRequest;
-import com.spotify.helios.transport.RequestDispatcher;
-import com.spotify.helios.transport.Response;
 
 import org.apache.http.conn.ssl.DefaultHostnameVerifier;
 import org.slf4j.Logger;
