+++ helios-client/src/main/java/com/spotify/helios/client/HeliosRequest.java
@@ -19,7 +19,7 @@
  * under the License.
  */
 
-package com.spotify.helios.transport;
+package com.spotify.helios.client;
 
 import com.google.common.collect.ImmutableMap;
 import com.google.common.collect.Lists;
